+++ packages/cv/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.2.3+1
+
+- Allow nullable for `CvField.withValue`, value being always set, even for null.
+- Fix `fillField` for generic `List` and `Map`
+
+## 0.2.2
+
+- Add `CvField.isNotNull`, `CvField.valueOrNull` (getter and setter) and `CvField.valueOrThrow` (getter and setter)
+
+## 0.2.1+1
+
+- Add cv() as extension on List<Map>.
+
 ## 0.2.0
 
 - Fail when child/children builder are missing